Add further functionality to line items in InvoiceForm component

+++ src/routes/(dashboard)/invoices/LineItemRow.svelte
@@ -1,36 +1,72 @@
 <script lang="ts">
-	import TrashIcon from "$lib/components/Icon/TrashIcon.svelte";
 	import { createEventDispatcher } from "svelte";
+	import TrashIcon from "$lib/components/Icon/TrashIcon.svelte";
+	import { twoDecimals, dollarsToCents } from "$lib/utils/moneyHelpers";
 
 	let dispatch = createEventDispatcher();
 
 	export let lineItem: LineItem;
+	export let canDelete = false;
+
+	let unitPrice = twoDecimals(lineItem.amount / lineItem.quantity);
+	let amount: string;
+
+	$: {
+		amount = twoDecimals(lineItem.quantity * Number(unitPrice));
+		lineItem.amount = dollarsToCents(Number(amount));
+	}
 </script>
 
 <div class="invoice-line-item border-b-2 border-fog py-2">
 	<div>
-		<input class="line-item" type="text" name="description" />
+		<input class="line-item" type="text" name="description" bind:value={lineItem.description} />
 	</div>
 
 	<div>
-		<input class="line-item text-right" type="number" name="unitPrice" step="0.01" min="0" />
+		<input
+			class="line-item text-right"
+			type="number"
+			name="unitPrice"
+			step="0.01"
+			min="0"
+			bind:value={unitPrice}
+			on:blur={() => {
+				unitPrice = twoDecimals(Number(unitPrice));
+			}}
+		/>
 	</div>
 
 	<div>
-		<input class="line-item text-center" type="number" name="quantity" min="0" />
+		<input
+			class="line-item text-center"
+			type="number"
+			name="quantity"
+			min="0"
+			bind:value={lineItem.quantity}
+		/>
 	</div>
 
 	<div>
-		<input class="line-item text-right" type="number" name="amount" step="0.01" min="0" />
+		<input
+			class="line-item text-right"
+			type="number"
+			name="amount"
+			step="0.01"
+			min="0"
+			bind:value={amount}
+			disabled
+		/>
 	</div>
 
 	<div>
-		<button
-			class="center h-10 w-10 text-pastelPurple hover:text-lavenderIndigo"
-			on:click|preventDefault={() => {
-				dispatch("removeLineItem", lineItem.id);
-			}}><TrashIcon /></button
-		>
+		{#if canDelete}
+			<button
+				class="center h-10 w-10 text-pastelPurple hover:text-lavenderIndigo"
+				on:click|preventDefault={() => {
+					dispatch("removeLineItem", lineItem.id);
+				}}><TrashIcon /></button
+			>
+		{/if}
 	</div>
 </div>
 
@@ -51,5 +87,10 @@
 	input[type="text"]:focus,
 	input[type="number"]:focus {
 		@apply border-solid border-lavenderIndigo outline-none;
+	}
+
+	input[type="text"]:disabled,
+	input[type="number"]:disabled {
+		@apply border-b-0 bg-transparent px-0;
 	}
 </style>
